+++ apps/faucet/package.json
@@ -14,9 +14,11 @@
     "@radix-ui/react-label": "^2.1.0",
     "@radix-ui/react-select": "^2.1.1",
     "@radix-ui/react-slot": "^1.1.0",
-    "@rainbow-me/rainbowkit": "2.x",
+    "@rainbow-me/rainbowkit": "2.1.2",
     "@tanstack/react-query": "^5.51.11",
-    "@wagmi/connectors": "5.0.26",
+    "@wagmi/connectors": "^5.1.1",
+    "app-contracts": "workspace:*",
+    "app-env": "workspace:*",
     "autoprefixer": "^10.4.19",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
@@ -24,13 +26,11 @@
     "postcss": "^8.4.39",
     "react": "18.x",
     "react-dom": "18.x",
-    "app-env": "workspace:*",
-    "app-contracts": "workspace:*",
     "tailwind-merge": "^2.4.0",
     "tailwindcss": "^3.4.6",
     "tailwindcss-animate": "^1.0.7",
     "viem": "latest",
-    "wagmi": "2.9.3"
+    "wagmi": "^2.12.1"
   },
   "devDependencies": {
     "@types/node": "latest",
