+++ apps/webapp/package.json
@@ -15,8 +15,7 @@
     "format": "prettier --write \"{app,lib,components}/**/*.{ts,tsx,mdx}\" --cache",
     "format:check": "prettier --check \"{app,lib,components}**/*.{ts,tsx,mdx}\" --cache",
     "translate": "bun run scripts/translate.ts",
-    "translate:category": "bun run scripts/translate-category.ts",
-    "vercel:install": "npm i -g node-gyp; bun install"
+    "translate:category": "bun run scripts/translate-category.ts"
   },
   "dependencies": {
     "@ai-sdk/anthropic": "^0.0.32",
@@ -41,7 +40,7 @@
     "@radix-ui/react-slot": "^1.1.0",
     "@radix-ui/react-tabs": "^1.1.0",
     "@radix-ui/react-tooltip": "^1.1.2",
-    "@rainbow-me/rainbowkit": "2.x",
+    "@rainbow-me/rainbowkit": "2.1.2",
     "@supabase/ssr": "^0.4.0",
     "@supabase/supabase-js": "^2.44.4",
     "@tailwindcss/typography": "0.5.13",
@@ -49,9 +48,12 @@
     "@vercel/analytics": "^1.3.1",
     "@vercel/og": "^0.6.2",
     "@vercel/speed-insights": "^1.0.12",
-    "@wagmi/connectors": "5.0.26",
+    "@wagmi/connectors": "^5.1.1",
     "@wharfkit/antelope": "^1.0.8",
     "ai": "^3.2.35",
+    "app-contracts": "workspace:*",
+    "app-env": "workspace:*",
+    "app-lib": "workspace:*",
     "axios": "^1.7.2",
     "camelize-ts": "^3.0.0",
     "class-variance-authority": "^0.7.0",
@@ -69,6 +71,7 @@
     "next": "14.2.5",
     "next-share": "^0.27.0",
     "next-themes": "^0.3.0",
+    "nookies": "^2.5.2",
     "pako": "^2.1.0",
     "pino-pretty": "^11.2.1",
     "react": "18.x",
@@ -86,16 +89,12 @@
     "resend": "^3.5.0",
     "server-only": "^0.0.1",
     "sharp": "^0.33.4",
-    "app-contracts": "workspace:*",
-    "app-env": "workspace:*",
-    "app-lib": "workspace:*",
     "uuid": "^10.0.0",
     "vconsole": "^3.15.1",
     "vercel": "^35.1.0",
     "viem": "latest",
-    "wagmi": "2.9.3",
-    "zod": "^3.23.8",
-    "nookies": "^2.5.2"
+    "wagmi": "^2.12.1",
+    "zod": "^3.23.8"
   },
   "devDependencies": {
     "@tailwindcss/typography": "^0.5.13",
