+++ packages/app-contracts/src/index.ts
@@ -1,4 +1,13 @@
-export * from "./dev";
-export * from "./prod";
-export * from "./types";
+import {devContracts} from "./dev"
+import {prodContracts} from "./prod"
+
+export * from './dev'
+export * from './prod'
+export * from "./types"
+
+export const appContracts = {
+  dev: devContracts,
+  prod: prodContracts
+}
+
