+++ packages/app-contracts/src/prod/tokens/usdt.ts
@@ -1,8 +1,8 @@
-import { TokenContractData } from "../types";
+import { EVMTokenContractData } from '../../types'
 import { erc20Abi } from 'abitype/abis'
+import { mainnet, arbitrum, avalanche, bsc, fantom, polygon, optimism, moonbeam, cronos, kava, metis, gnosis, aurora, harmonyOne } from 'viem/chains'
 
-
-export const usdtContracts: TokenContractData[] = [
+export const usdtContracts: EVMTokenContractData[] = [
   {
     address: "0xdAC17F958D2ee523a2206206994597C13D831ec7",
     name: "USDT",
@@ -13,6 +13,7 @@ export const usdtContracts: TokenContractData[] = [
     chainType: "evm",
     chainName: "Ethereum",
     abi: erc20Abi,
+    chain: mainnet,
   },
   {
     address: "0xFD086BC7CD5C481DCC9C85E0C3AC0A4A5404C120",
@@ -24,6 +25,7 @@ export const usdtContracts: TokenContractData[] = [
     chainType: "evm",
     chainName: "Arbitrum",
     abi: erc20Abi,
+    chain: arbitrum,
   },
   {
     address: "0xc7198437980c041c805A1EDcbA50c1Ce5db95118",
@@ -35,6 +37,7 @@ export const usdtContracts: TokenContractData[] = [
     chainType: "evm",
     chainName: "Avalanche",
     abi: erc20Abi,
+    chain: avalanche,
   },
   {
     address: "0x55d398326f99059fF775485246999027B3197955",
@@ -46,6 +49,7 @@ export const usdtContracts: TokenContractData[] = [
     chainType: "evm",
     chainName: "Binance Smart Chain",
     abi: erc20Abi,
+    chain: bsc,
   },
   {
     address: "0x049d68029688eAbF473097a2fC38ef61633A3C7A",
@@ -57,6 +61,7 @@ export const usdtContracts: TokenContractData[] = [
     chainType: "evm",
     chainName: "Fantom",
     abi: erc20Abi,
+    chain: fantom,
   },
   {
     address: "0x3813e82e6f7098b9583FC0F33a962D02018B6803",
@@ -68,17 +73,7 @@ export const usdtContracts: TokenContractData[] = [
     chainType: "evm",
     chainName: "Polygon",
     abi: erc20Abi,
-  },
-  {
-    address: "TD9iWjeY9yLbzgpqv9E5uFZy9vNExS6EN5",
-    name: "USDT",
-    symbol: "USDT",
-    decimals: 6,
-    indexFromBlock: 0,
-    chainId: 1, // Tron Mainnet (TRC-20)
-    chainType: "evm",
-    chainName: "Tron",
-    abi: erc20Abi,
+    chain: polygon,
   },
   {
     address: "0x7F5c764cBc14f9669B88837ca1490cCa17c31607",
@@ -90,6 +85,7 @@ export const usdtContracts: TokenContractData[] = [
     chainType: "evm",
     chainName: "Optimism",
     abi: erc20Abi,
+    chain: optimism,
   },
   {
     address: "0x224E64ec1BDce3870a6a6c777eDd450454068FEC",
@@ -101,6 +97,7 @@ export const usdtContracts: TokenContractData[] = [
     chainType: "evm",
     chainName: "Harmony",
     abi: erc20Abi,
+    chain: harmonyOne,
   },
   {
     address: "0x8e76f63A2dC3F3bE6f00e3F4a1ddC133D52f4866",
@@ -112,6 +109,7 @@ export const usdtContracts: TokenContractData[] = [
     chainType: "evm",
     chainName: "Moonbeam",
     abi: erc20Abi,
+    chain: moonbeam,
   },
   {
     address: "0x6175F5a135dF205eD2E6FCab46dEc8df3feBf60a",
@@ -123,6 +121,7 @@ export const usdtContracts: TokenContractData[] = [
     chainType: "evm",
     chainName: "Cronos",
     abi: erc20Abi,
+    chain: cronos,
   },
   {
     address: "0x33bD5cD41aE43Fd478C14d0FF16B23E92cb9d015",
@@ -134,6 +133,7 @@ export const usdtContracts: TokenContractData[] = [
     chainType: "evm",
     chainName: "Kava",
     abi: erc20Abi,
+    chain: kava,
   },
   {
     address: "0x29db9dAF5b1D73f5A3265B5Edb2Dd9d2b99D6A3A",
@@ -145,6 +145,7 @@ export const usdtContracts: TokenContractData[] = [
     chainType: "evm",
     chainName: "Metis",
     abi: erc20Abi,
+    chain: metis,
   },
   {
     address: "0x4ecb6e9aEa4C7D2b16B65F6375A7b88d14524eC9",
@@ -156,6 +157,7 @@ export const usdtContracts: TokenContractData[] = [
     chainType: "evm",
     chainName: "Gnosis",
     abi: erc20Abi,
+    chain: gnosis,
   },
   {
     address: "0x4988a896b1227218e4A686fdE5EabdcAbd91571f",
@@ -167,5 +169,6 @@ export const usdtContracts: TokenContractData[] = [
     chainType: "evm",
     chainName: "Aurora",
     abi: erc20Abi,
+    chain: aurora,
   },
 ]
