--- packages/app-contracts/src/prod/usdt.ts
@@ -1,171 +1,0 @@
-import { TokenContractData } from "../types";
-import { erc20Abi } from 'abitype/abis'
-
-
-export const usdtContracts: TokenContractData[] = [
-  {
-    address: "0xdAC17F958D2ee523a2206206994597C13D831ec7",
-    name: "USDT",
-    symbol: "USDT",
-    decimals: 6,
-    indexFromBlock: 0,
-    chainId: 1, // Ethereum Mainnet
-    chainType: "evm",
-    chainName: "Ethereum",
-    abi: erc20Abi,
-  },
-  {
-    address: "0xFD086BC7CD5C481DCC9C85E0C3AC0A4A5404C120",
-    name: "USDT",
-    symbol: "USDT",
-    decimals: 6,
-    indexFromBlock: 0,
-    chainId: 42161, // Arbitrum Mainnet
-    chainType: "evm",
-    chainName: "Arbitrum",
-    abi: erc20Abi,
-  },
-  {
-    address: "0xc7198437980c041c805A1EDcbA50c1Ce5db95118",
-    name: "USDT",
-    symbol: "USDT",
-    decimals: 6,
-    indexFromBlock: 0,
-    chainId: 43114, // Avalanche Mainnet
-    chainType: "evm",
-    chainName: "Avalanche",
-    abi: erc20Abi,
-  },
-  {
-    address: "0x55d398326f99059fF775485246999027B3197955",
-    name: "USDT",
-    symbol: "USDT",
-    decimals: 6,
-    indexFromBlock: 0,
-    chainId: 56, // Binance Smart Chain Mainnet
-    chainType: "evm",
-    chainName: "Binance Smart Chain",
-    abi: erc20Abi,
-  },
-  {
-    address: "0x049d68029688eAbF473097a2fC38ef61633A3C7A",
-    name: "USDT",
-    symbol: "USDT",
-    decimals: 6,
-    indexFromBlock: 0,
-    chainId: 250, // Fantom Mainnet
-    chainType: "evm",
-    chainName: "Fantom",
-    abi: erc20Abi,
-  },
-  {
-    address: "0x3813e82e6f7098b9583FC0F33a962D02018B6803",
-    name: "USDT",
-    symbol: "USDT",
-    decimals: 6,
-    indexFromBlock: 0,
-    chainId: 137, // Polygon Mainnet
-    chainType: "evm",
-    chainName: "Polygon",
-    abi: erc20Abi,
-  },
-  {
-    address: "TD9iWjeY9yLbzgpqv9E5uFZy9vNExS6EN5",
-    name: "USDT",
-    symbol: "USDT",
-    decimals: 6,
-    indexFromBlock: 0,
-    chainId: 1, // Tron Mainnet (TRC-20)
-    chainType: "evm",
-    chainName: "Tron",
-    abi: erc20Abi,
-  },
-  {
-    address: "0x7F5c764cBc14f9669B88837ca1490cCa17c31607",
-    name: "USDT",
-    symbol: "USDT",
-    decimals: 6,
-    indexFromBlock: 0,
-    chainId: 10, // Optimism Mainnet
-    chainType: "evm",
-    chainName: "Optimism",
-    abi: erc20Abi,
-  },
-  {
-    address: "0x224E64ec1BDce3870a6a6c777eDd450454068FEC",
-    name: "USDT",
-    symbol: "USDT",
-    decimals: 6,
-    indexFromBlock: 0,
-    chainId: 1666600000, // Harmony Mainnet
-    chainType: "evm",
-    chainName: "Harmony",
-    abi: erc20Abi,
-  },
-  {
-    address: "0x8e76f63A2dC3F3bE6f00e3F4a1ddC133D52f4866",
-    name: "USDT",
-    symbol: "USDT",
-    decimals: 6,
-    indexFromBlock: 0,
-    chainId: 1284, // Moonbeam Mainnet
-    chainType: "evm",
-    chainName: "Moonbeam",
-    abi: erc20Abi,
-  },
-  {
-    address: "0x6175F5a135dF205eD2E6FCab46dEc8df3feBf60a",
-    name: "USDT",
-    symbol: "USDT",
-    decimals: 6,
-    indexFromBlock: 0,
-    chainId: 25, // Cronos Mainnet
-    chainType: "evm",
-    chainName: "Cronos",
-    abi: erc20Abi,
-  },
-  {
-    address: "0x33bD5cD41aE43Fd478C14d0FF16B23E92cb9d015",
-    name: "USDT",
-    symbol: "USDT",
-    decimals: 6,
-    indexFromBlock: 0,
-    chainId: 2222, // Kava Mainnet
-    chainType: "evm",
-    chainName: "Kava",
-    abi: erc20Abi,
-  },
-  {
-    address: "0x29db9dAF5b1D73f5A3265B5Edb2Dd9d2b99D6A3A",
-    name: "USDT",
-    symbol: "USDT",
-    decimals: 6,
-    indexFromBlock: 0,
-    chainId: 1088, // Metis Mainnet
-    chainType: "evm",
-    chainName: "Metis",
-    abi: erc20Abi,
-  },
-  {
-    address: "0x4ecb6e9aEa4C7D2b16B65F6375A7b88d14524eC9",
-    name: "USDT",
-    symbol: "USDT",
-    decimals: 6,
-    indexFromBlock: 0,
-    chainId: 100, // Gnosis Mainnet
-    chainType: "evm",
-    chainName: "Gnosis",
-    abi: erc20Abi,
-  },
-  {
-    address: "0x4988a896b1227218e4A686fdE5EabdcAbd91571f",
-    name: "USDT",
-    symbol: "USDT",
-    decimals: 6,
-    indexFromBlock: 0,
-    chainId: 1313161554, // Aurora Mainnet
-    chainType: "evm",
-    chainName: "Aurora",
-    abi: erc20Abi,
-  },
-]
